backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20

# Authentication
bcrypt==4.2.1
PyJWT==2.10.1

# Data Validation
pydantic==2.10.3
pydantic[email]==2.10.3

# ML - Document Classification
sentence-transformers==3.3.1
torch>=2.6.0
scikit-learn>=1.6.0
numpy>=2.2.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# LLM Integration - Google Gemini
google-generativeai>=0.8.0

# Environment Variables
python-dotenv==1.0.1

# Rate Limiting
slowapi>=0.1.9